Refactor gallery and settings screens; implement cropping functionality
- Replaced the temporary re-export of RedoCropScreen with a complete implementation that includes image cropping functionality, hint display, and image processing using ChatGPT service.
- Updated SettingsScreen to provide options for changing and resetting the import directory, along with cloud account sign-in status and last sync time display.
- Adjusted image_gallery_test to import the correct image gallery screen for testing.
