+++ REDO_ENHANCEMENTS.md
@@ -1,0 +1,70 @@
+# Redo Functionality Enhancements
+
+## Overview
+Enhanced the redo OCR functionality with two key improvements addressing user experience and mobile usability.
+
+## Changes Made
+
+### 1. Gesture Shortcut - Double-Tap
+**Problem**: Three-finger gesture was impractical for mobile devices
+**Solution**: Replaced with double-tap gesture
+
+- **Gesture**: Double-tap on any image
+- **Mobile-friendly**: Natural single-hand operation
+- **Non-conflicting**: Doesn't interfere with existing gestures:
+  - Single tap → Show details dialog
+  - Long press → Select/deselect image
+  - PhotoView gestures → Pinch to zoom, drag to pan
+- **Feedback**: Light haptic feedback when triggered
+- **Updated tooltip**: "Double-tap on image for quick redo"
+
+### 2. Interactive Crop Screen
+**Problem**: Crop screen wasn't interactive and couldn't resize crop area
+**Solution**: Complete redesign with PhotoView integration and resizable crop
+
+#### Features:
+- **Full zoom capability**: Pinch to zoom in/out up to 3x
+- **Pan support**: Drag to move image around
+- **Resizable crop area**: Drag corners to resize selection
+- **Movable crop area**: Drag center to reposition
+- **Visual enhancements**:
+  - Rule of thirds grid lines for better composition
+  - Corner handles with visual feedback
+  - Active handle highlighting (blue when resizing)
+  - Semi-transparent overlay
+  - Improved button design
+
+#### User Experience:
+- **Better instructions**: Clear hints with emoji icons
+- **Dismissible hints**: "Got it!" button to hide instructions
+- **Improved controls**: Icon buttons with better styling
+- **Error handling**: Better error messages and state management
+- **Memory management**: Proper image disposal to prevent leaks
+
+## Technical Implementation
+
+### Gesture Detection
+```dart
+onDoubleTap: () {
+  _redoTextExtraction();
+  HapticFeedback.lightImpact();
+},
+```
+
+### Interactive Crop
+- **PhotoView integration**: Full zoom and pan capabilities
+- **Custom painter**: Enhanced crop overlay with grid and handles
+- **Gesture handling**: Pan detection for move/resize operations
+- **Coordinate transformation**: Accurate crop area mapping from screen to image coordinates
+- **Image processing**: Direct image manipulation without RepaintBoundary
+
+## Files Modified
+- `lib/screens/gallery/image_gallery_screen.dart`
+- `lib/screens/gallery/redo_crop_screen.dart`
+
+## Benefits
+1. **Mobile-optimized**: Double-tap is natural on touchscreens
+2. **Precise cropping**: Zoom in for accurate text selection
+3. **Flexible selection**: Resizable crop area for any text size
+4. **Better UX**: Clear visual feedback and instructions
+5. **Improved accuracy**: Higher precision in OCR text selection
