+++ lib/screens/gallery/image_gallery_screen.dart
@@ -1,4 +1,1847 @@
-// Moved from experimental/2attempt/imageGalleryScreen.dart
-// This file reorganizes the New UI gallery under lib/screens/gallery.
+import 'dart:io';
+import 'dart:async';
 
-export '../../experimental/2attempt/imageGalleryScreen.dart';
+import 'package:PhotoWordFind/models/contactEntry.dart';
+import 'package:PhotoWordFind/social_icons.dart';
+import 'package:PhotoWordFind/utils/storage_utils.dart';
+import 'package:PhotoWordFind/services/search_service.dart';
+import 'package:PhotoWordFind/services/chat_gpt_service.dart';
+import 'package:wechat_assets_picker/wechat_assets_picker.dart';
+import 'package:PhotoWordFind/utils/chatgpt_post_utils.dart';
+import 'package:photo_manager/photo_manager.dart';
+import 'package:file_picker/file_picker.dart';
+import 'package:path/path.dart' as path;
+import 'package:flutter/material.dart';
+import 'package:flutter/gestures.dart';
+import 'package:photo_view/photo_view.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+import 'package:flutter/services.dart';
+import 'package:url_launcher/url_launcher.dart';
+import 'package:PhotoWordFind/widgets/note_dialog.dart';
+import 'package:PhotoWordFind/widgets/confirmation_dialog.dart';
+import 'package:PhotoWordFind/screens/settings/settings_screen.dart';
+import 'package:PhotoWordFind/screens/gallery/redo_crop_screen.dart';
+import 'package:PhotoWordFind/utils/cloud_utils.dart';
+import 'package:permission_handler/permission_handler.dart';
+import 'package:intl/intl.dart';
+import 'package:PhotoWordFind/main.dart' show UiMode; // for UI mode switching
+
+final PageController _pageController =
+		PageController(viewportFraction: 0.8); // Gives a gallery feel
+
+class ImageGalleryScreen extends StatefulWidget {
+	@override
+	_ImageGalleryScreenState createState() => _ImageGalleryScreenState();
+}
+
+class _ImageGalleryScreenState extends State<ImageGalleryScreen>
+		with TickerProviderStateMixin {
+	String searchQuery = '';
+	String selectedSortOption = 'Name'; // Default value from the list
+	List<String> sortOptions = [
+		'Name',
+		'Date found',
+		'Size',
+		'Snap Added Date',
+		'Instagram Added Date',
+		'Added on Snapchat',
+		'Added on Instagram'
+	]; // Add your sort options here
+	String selectedState = 'All';
+	List<String> states = ['All'];
+	List<ContactEntry> images = [];
+	List<ContactEntry> allImages = [];
+	List<String> selectedImages = [];
+
+	int currentIndex = 0;
+	static const String _lastStateKey = 'last_selected_state';
+	static const String _importDirKey = 'import_directory';
+	String? _importDirPath;
+
+	// State variable to track the selected sort order
+	bool isAscending = true; // Default sorting order
+
+	// Add a setting to control which loading method to use
+	bool useJsonFileForLoading = false; // Set to true to load from JSON file
+
+	bool _controlsExpanded = true; // Tracks whether the controls are minimized
+
+	bool _isInitializing = true; // Tracks if app is still initializing
+	String? _initializationError; // Stores any initialization error
+
+	/// Key for the root [Navigator] so dialogs can use a context that
+	/// has navigation and localization available.
+	final GlobalKey<NavigatorState> _navigatorKey = GlobalKey<NavigatorState>();
+
+	// Track sign-out operation (removed icon usage, state no longer needed)
+	// bool _signingOut = false;
+	// String? _signOutMessage;
+	// New sync state
+	bool _syncing = false;
+	DateTime? _lastSyncTime;
+	// String? _lastSyncError; // no longer surfaced
+
+	@override
+	void initState() {
+		super.initState();
+		// Attach progress callback for CloudUtils (scoped to this screen)
+		CloudUtils.progressCallback = ({double? value, String? message, bool done = false, bool error = false}) {
+			if (!mounted) return;
+			// Previously updated sign-out UI; now no-op.
+		};
+		_initializeApp();
+	}
+
+	Future requestPermissions() async {
+		var status = await Permission.manageExternalStorage.status;
+		if (!status.isGranted) {
+			await Permission.manageExternalStorage.request();
+		}
+	}
+
+	/// Initializes the app in sequential order:
+	/// 1. Sign-in first
+	/// 2. Load images after sign-in is complete
+	/// 3. Load import directory
+	Future<void> _initializeApp() async {
+		try {
+			await requestPermissions();
+			setState(() {
+				_isInitializing = true;
+				_initializationError = null;
+			});
+
+			// Step 1: Ensure user is signed in first
+			final signedIn = await _ensureSignedIn();
+
+			if (!signedIn) {
+				setState(() {
+					_initializationError =
+							'Sign-in failed. Some features may not be available.';
+				});
+			}
+
+			// Step 2: Load images only after sign-in is complete
+			if (useJsonFileForLoading) {
+				await _loadImagesFromJsonFile();
+			} else {
+				await _loadImagesFromPreferences();
+			}
+
+			// Step 3: Load import directory
+			await _loadImportDirectory();
+
+			setState(() {
+				_isInitializing = false;
+			});
+		} catch (e) {
+			setState(() {
+				_isInitializing = false;
+				_initializationError = 'Initialization failed: $e';
+			});
+		}
+	}
+
+	Future<bool> _ensureSignedIn() async {
+		// First check if we're already signed in (from main app initialization)
+		bool signed = await CloudUtils.isSignedin();
+		if (!signed) {
+			// Only attempt sign-in if not already signed in
+			signed = await CloudUtils.firstSignIn();
+		}
+		return signed;
+	}
+
+	Future<void> _forceSync({bool fromPull = false}) async {
+		if (_syncing) return; // Prevent concurrent syncs
+		final ctx = _navigatorKey.currentContext ?? context;
+		final messenger = ScaffoldMessenger.of(ctx);
+
+		setState(() {
+			_syncing = true;
+		});
+
+		void show(String msg, {Duration duration = const Duration(seconds: 2)}) {
+			if (fromPull) return;
+			messenger.showSnackBar(SnackBar(content: Text(msg), duration: duration));
+		}
+
+		try {
+			if (!await CloudUtils.isSignedin()) {
+				show('Not signed in');
+				return;
+			}
+			if (!await CloudUtils.isConnected()) {
+				show('No internet connection');
+				return;
+			}
+			if (!fromPull) {
+				messenger.hideCurrentSnackBar();
+				messenger.showSnackBar(const SnackBar(
+					content: Row(
+						children: [
+							SizedBox(width: 18, height: 18, child: CircularProgressIndicator(strokeWidth: 2)),
+							SizedBox(width: 12),
+							Text('Syncing…'),
+						],
+					),
+					duration: Duration(minutes: 1),
+				));
+			}
+			await CloudUtils.updateCloudJson().timeout(const Duration(seconds: 45));
+			_lastSyncTime = DateTime.now();
+			if (!fromPull) {
+				messenger.hideCurrentSnackBar();
+				show('Sync complete (${DateFormat.Hm().format(_lastSyncTime!)})');
+			}
+		} catch (e) {
+			if (!fromPull) {
+				messenger.hideCurrentSnackBar();
+				show('Sync failed: $e');
+			}
+		} finally {
+			if (mounted) setState(() => _syncing = false);
+		}
+	}
+
+	Future<void> _toggleSignInOut() async {
+		final ctx = _navigatorKey.currentContext ?? context;
+		bool signed = await CloudUtils.isSignedin();
+		if (!signed) {
+			setState(() {
+				_isInitializing = true;
+			});
+			try {
+				await CloudUtils.firstSignIn();
+			} finally {
+				if (!mounted) return;
+				setState(() {
+					_isInitializing = false;
+				});
+			}
+		} else {
+			final confirm = await showDialog<bool>(
+				context: ctx,
+				builder: (_) => ConfirmationDialog(message: 'Sign out?'),
+			);
+			if (confirm == true) {
+				setState(() {
+					// _signingOut = true; // removed
+					// _signOutMessage = 'Signing out...'; // removed
+				});
+				try {
+					await CloudUtils.signOut();
+				} catch (_) {
+					// error message already captured via callback
+				}
+			}
+		}
+		if (!mounted) return;
+		setState(() {
+			// Trigger UI update to reflect new sign-in state
+		});
+	}
+
+	Future<void> _loadImagesFromPreferences() async {
+		// Read the image path map from the JSON file (simulating a separate storage location)
+		List<ContactEntry> loadedImages = [];
+
+		// Fetch each contact directly from Hive using the keys list to avoid
+		// reading the entire box twice.
+		for (final identifier in StorageUtils.getKeys()) {
+			final contactEntry = await StorageUtils.get(identifier);
+			if (contactEntry != null) {
+				loadedImages.add(contactEntry);
+			}
+		}
+		allImages = loadedImages;
+		_updateStates(allImages);
+		await _restoreLastSelectedState();
+		await _applyFiltersAndSort();
+	}
+
+	Future<void> _loadImagesFromJsonFile() async {
+		// Read the image path map from the JSON file
+		final Map<String, dynamic> fileMap = await StorageUtils.readJson();
+		List<ContactEntry> loadedImages = [];
+		for (final entry in fileMap.entries) {
+			final identifier = entry.key;
+			final value = entry.value;
+			// If the value is just a path string, create a minimal ContactEntry
+			if (value is String) {
+				loadedImages.add(ContactEntry(
+					identifier: identifier,
+					imagePath: value,
+					dateFound: File(value).existsSync()
+							? File(value).lastModifiedSync()
+							: DateTime.now(),
+					json: {'imagePath': value},
+				));
+			} else if (value is Map<String, dynamic>) {
+				// If the value is a full contact JSON, use fromJson
+				final imagePath = value['imagePath'] ?? '';
+				loadedImages.add(ContactEntry.fromJson(
+					identifier,
+					imagePath,
+					value,
+				));
+			}
+		}
+		allImages = loadedImages;
+		_updateStates(allImages);
+		await _restoreLastSelectedState();
+		await _applyFiltersAndSort();
+	}
+
+		@override
+		Widget build(BuildContext context) {
+			// No nested MaterialApp. Use the app's root MaterialApp.
+			final navContext = context;
+			return Scaffold(
+						appBar: AppBar(
+							title: const Text('Image Gallery'),
+							actions: [
+								IconButton(
+									tooltip: 'Switch to Legacy UI',
+									icon: const Icon(Icons.swap_horiz),
+									onPressed: () async {
+										final proceed = await showDialog<bool>(
+											context: navContext,
+											builder: (dialogCtx) => AlertDialog(
+												title: const Text('Confirm UI Switch'),
+												content: const Text('Switch to legacy interface?'),
+												actions: [
+													TextButton(
+														onPressed: () => Navigator.pop(dialogCtx, false),
+														child: const Text('Cancel'),
+													),
+													TextButton(
+														onPressed: () => Navigator.pop(dialogCtx, true),
+														child: const Text('Switch'),
+													),
+												],
+											),
+										);
+										if (proceed == true) {
+											await UiMode.switchTo(false);
+										}
+									},
+								),
+								if (_isInitializing)
+									const Padding(
+										padding: EdgeInsets.symmetric(horizontal: 16.0),
+										child: SizedBox(
+											width: 24,
+											height: 24,
+											child: CircularProgressIndicator(strokeWidth: 2),
+										),
+									)
+								else
+									// Persistent auth status icon (green if signed in, red if failed last attempt)
+									FutureBuilder<bool>(
+										future: CloudUtils.isSignedin(),
+										builder: (ctx, snap) {
+											final signed = snap.data == true;
+											return Padding(
+												padding: const EdgeInsets.symmetric(horizontal: 4.0),
+												child: Icon(
+													signed ? Icons.cloud_done : Icons.cloud_off,
+													color: signed ? Colors.green : Colors.redAccent,
+												),
+											);
+										},
+									),
+								if (!_isInitializing)
+									IconButton(
+										icon: const Icon(Icons.settings),
+										tooltip: 'Settings',
+										onPressed: () {
+											Navigator.of(navContext).push(
+												MaterialPageRoute(
+													builder: (_) => SettingsScreen(
+														onResetImportDir: _resetImportDir,
+														onChangeImportDir: _changeImportDir,
+														onRequestSignIn: () async {
+															final ok = await _ensureSignedIn();
+															if (!ok) {
+																setState(() {
+																	_initializationError = 'Sign-in failed';
+																});
+															}
+															return ok;
+														},
+														onRequestSignOut: () async {
+															try {
+																await CloudUtils.signOut();
+															} catch (_) {}
+															setState(() {});
+														},
+														lastSyncTime: _lastSyncTime,
+														syncing: _syncing,
+														signInFailed: _initializationError != null,
+													),
+												),
+											);
+										},
+									),
+							],
+						),
+						body: _isInitializing
+								? Center(
+										child: Column(
+											mainAxisAlignment: MainAxisAlignment.center,
+											children: [
+												const CircularProgressIndicator(),
+												const SizedBox(height: 16),
+												const Text('Signing in and loading images...'),
+												if (_initializationError != null) ...[
+													const SizedBox(height: 16),
+													Padding(
+														padding:
+																const EdgeInsets.symmetric(horizontal: 16.0),
+														child: Text(
+															_initializationError!,
+															style: const TextStyle(color: Colors.orange),
+															textAlign: TextAlign.center,
+														),
+													),
+												],
+											],
+										),
+									)
+								: LayoutBuilder(
+										builder: (context, constraints) {
+											final screenHeight = constraints.maxHeight;
+											return Column(
+												children: [
+													_buildControls(),
+													Expanded(
+														child: ImageGallery(
+															images: images,
+															selectedImages: selectedImages,
+															sortOption: selectedSortOption,
+															onImageSelected: (String id) {
+																setState(() {
+																	if (selectedImages.contains(id)) {
+																		selectedImages.remove(id);
+																	} else {
+																		selectedImages.add(id);
+																	}
+																});
+															},
+															onMenuOptionSelected: _onMenuOptionSelected,
+															galleryHeight: screenHeight,
+															onPageChanged: (idx) =>
+																	setState(() => currentIndex = idx),
+															currentIndex: currentIndex,
+														),
+													),
+												],
+											);
+										},
+									),
+						floatingActionButton: _isInitializing
+								? null // Hide FAB during initialization
+								: selectedImages.isNotEmpty
+										? FloatingActionButton(
+												onPressed: () => _onMenuOptionSelected('', 'move'),
+												child: Icon(Icons.move_to_inbox),
+											)
+										: FloatingActionButton(
+												onPressed: () => _importImages(navContext),
+												tooltip: 'Import Images',
+												child: Icon(Icons.add),
+											),
+						persistentFooterButtons: [
+							SingleChildScrollView(
+								scrollDirection: Axis.horizontal,
+								child: Container(
+									width: MediaQuery.of(navContext).size.width * 1.20,
+									child: Row(
+										mainAxisAlignment: MainAxisAlignment.spaceBetween,
+										children: [
+											SocialIcon.snapchatIconButton!,
+											Spacer(),
+											SocialIcon.galleryIconButton!,
+											Spacer(),
+											SocialIcon.bumbleIconButton!,
+											Spacer(),
+											FloatingActionButton(
+												heroTag: null,
+												tooltip: 'Change current directory',
+												onPressed: _changeImportDir,
+												child: Icon(Icons.drive_folder_upload),
+											),
+											Spacer(),
+											SocialIcon.instagramIconButton!,
+											Spacer(),
+											SocialIcon.discordIconButton!,
+											Spacer(),
+											SocialIcon.kikIconButton!,
+										],
+									),
+								),
+							),
+									],
+							);
+				}
+
+	Widget _buildControls() {
+		// Pull-to-refresh limited to this controls area only
+		return RefreshIndicator(
+			onRefresh: () => _forceSync(fromPull: true),
+			displacement: 56,
+			child: SingleChildScrollView(
+				physics: const AlwaysScrollableScrollPhysics(),
+				padding: EdgeInsets.zero,
+				child: AnimatedSize(
+					duration: const Duration(milliseconds: 300),
+					child: _controlsExpanded
+							? _buildExpandedControls()
+							: _buildMinimizedControls(),
+				),
+			),
+		);
+	}
+
+	Widget _buildExpandedControls() {
+		return LayoutBuilder(builder: (context, constraints) {
+			final isWide = constraints.maxWidth > 500;
+			final children = <Widget>[
+				Expanded(
+					child: DropdownButtonFormField<String>(
+						value: states.contains(selectedState) ? selectedState : null,
+						isExpanded: true,
+						decoration: InputDecoration(
+							labelText: 'State',
+							border: const OutlineInputBorder(),
+							isDense: true,
+						),
+						items: states
+								.map((dir) => DropdownMenuItem<String>(
+											value: dir,
+											child: Text(dir),
+										))
+								.toList(),
+						onChanged: (value) async {
+							selectedState = value!;
+							await _saveLastSelectedState(selectedState);
+							await _filterImages();
+						},
+					),
+				),
+				const SizedBox(width: 8),
+				Expanded(
+					flex: 2,
+					child: TextField(
+						decoration: InputDecoration(
+							hintText: 'Search',
+							prefixIcon: const Icon(Icons.search),
+							border: const OutlineInputBorder(),
+							isDense: true,
+						),
+						onChanged: (value) async {
+							searchQuery = value;
+							await _filterImages();
+						},
+					),
+				),
+				const SizedBox(width: 8),
+				Expanded(
+					child: DropdownButtonFormField<String>(
+						value: selectedSortOption,
+						isExpanded: true,
+						decoration: InputDecoration(
+							labelText: 'Sort by',
+							border: const OutlineInputBorder(),
+							isDense: true,
+						),
+						items: sortOptions
+								.map((option) => DropdownMenuItem<String>(
+											value: option,
+											child: Text(option),
+										))
+								.toList(),
+						onChanged: (value) async {
+							selectedSortOption = value!;
+							await _applyFiltersAndSort();
+						},
+					),
+				),
+				const SizedBox(width: 12),
+				_buildOrderToggle(),
+			];
+
+			Widget content = isWide
+					? Row(children: children)
+					: Column(
+							crossAxisAlignment: CrossAxisAlignment.stretch,
+							children: [
+								Row(children: children.sublist(0, 3)),
+								const SizedBox(height: 12),
+								Row(children: children.sublist(3)),
+							],
+						);
+
+			return Card(
+				margin: const EdgeInsets.all(12),
+				elevation: 4,
+				shape: RoundedRectangleBorder(
+					borderRadius: BorderRadius.circular(16),
+				),
+				child: Padding(
+					padding: const EdgeInsets.all(12.0),
+					child: Stack(
+						clipBehavior: Clip.none,
+						children: [
+							Padding(
+								padding: const EdgeInsets.only(top: 8.0, right: 8.0),
+								child: content,
+							),
+							Positioned(
+								top: 0,
+								right: 0,
+								child: IconButton(
+									icon: const Icon(Icons.expand_less),
+									onPressed: () {
+										setState(() => _controlsExpanded = false);
+									},
+								),
+							),
+						],
+					),
+				),
+			);
+		});
+	}
+
+	Widget _buildMinimizedControls() {
+		return Card(
+			margin: const EdgeInsets.all(12),
+			elevation: 4,
+			shape: RoundedRectangleBorder(
+				borderRadius: BorderRadius.circular(16),
+			),
+			child: Padding(
+				padding: const EdgeInsets.all(12.0),
+				child: Row(
+					children: [
+						Expanded(
+							child: DropdownButtonFormField<String>(
+								value: sortOptions.contains(selectedSortOption)
+										? selectedSortOption
+										: null,
+								isExpanded: true,
+								decoration: const InputDecoration(
+									labelText: 'Sort by',
+									border: OutlineInputBorder(),
+									isDense: true,
+								),
+								items: sortOptions
+										.map((option) => DropdownMenuItem<String>(
+													value: option,
+													child: Text(option),
+												))
+										.toList(),
+								onChanged: (value) async {
+									selectedSortOption = value!;
+									await _applyFiltersAndSort();
+								},
+							),
+						),
+						const SizedBox(width: 8),
+						IconButton(
+							icon: const Icon(Icons.expand_more),
+							onPressed: () {
+								setState(() => _controlsExpanded = true);
+							},
+						),
+					],
+				),
+			),
+		);
+	}
+
+	Widget _buildOrderToggle() {
+		return Padding(
+			padding: const EdgeInsets.symmetric(horizontal: 4.0),
+			child: Ink(
+				decoration: BoxDecoration(
+					color: Colors.blue.withOpacity(0.1),
+					shape: BoxShape.circle,
+				),
+				child: IconButton(
+					icon: AnimatedRotation(
+						turns: isAscending ? 0 : 0.5,
+						duration: const Duration(milliseconds: 200),
+						child: const Icon(Icons.arrow_upward, color: Colors.blue),
+					),
+					onPressed: () async {
+						isAscending = !isAscending;
+						await _applyFiltersAndSort();
+					},
+				),
+			),
+		);
+	}
+
+	Future<void> _filterImages() async {
+		await _applyFiltersAndSort();
+	}
+
+	void _updateStates(List<ContactEntry> imgs) {
+		final tags = <String>{};
+		for (final img in imgs) {
+			if (img.state != null) tags.add(img.state!);
+		}
+		final sorted = tags.toList()..sort();
+		setState(() {
+			states = ['All', ...sorted];
+			if (!states.contains(selectedState)) {
+				selectedState = 'All';
+			}
+		});
+	}
+
+	Future<void> _restoreLastSelectedState() async {
+		final prefs = await SharedPreferences.getInstance();
+		final last = prefs.getString(_lastStateKey);
+		if (last != null && states.contains(last)) {
+			selectedState = last;
+		}
+	}
+
+	Future<void> _saveLastSelectedState(String value) async {
+		final prefs = await SharedPreferences.getInstance();
+		await prefs.setString(_lastStateKey, value);
+	}
+
+	Future<void> _loadImportDirectory() async {
+		final prefs = await SharedPreferences.getInstance();
+		_importDirPath = prefs.getString(_importDirKey);
+	}
+
+	Future<void> _saveImportDirectory(String path) async {
+		final prefs = await SharedPreferences.getInstance();
+		await prefs.setString(_importDirKey, path);
+	}
+
+	Future<void> _changeImportDir() async {
+		final dir = await FilePicker.platform.getDirectoryPath();
+		if (dir != null) {
+			setState(() {
+				_importDirPath = dir;
+			});
+			await _saveImportDirectory(dir);
+		}
+	}
+
+	Future<void> _resetImportDir() async {
+		final prefs = await SharedPreferences.getInstance();
+		await prefs.remove(_importDirKey);
+		setState(() {
+			_importDirPath = null;
+		});
+	}
+
+	Future<AssetPathEntity?> _getImportAlbum() async {
+		if (_importDirPath == null) return null;
+		final paths = await PhotoManager.getAssetPathList(
+			type: RequestType.image,
+			hasAll: true,
+		);
+		final dirName = path.basename(_importDirPath!);
+		for (final p in paths) {
+			if (p.name == dirName) {
+				return p;
+			}
+		}
+		return null;
+	}
+
+	/// Build a filter that restricts the picker to [_importDirPath].
+	Future<PMFilter?> _buildImportFilter() async {
+		if (_importDirPath == null) return null;
+		final album = await _getImportAlbum();
+		if (album == null) return null;
+		if (Platform.isAndroid) {
+			final id = album.id.replaceAll("'", "''");
+			return CustomFilter.sql(
+				where: "${CustomColumns.android.bucketId} = '$id'",
+			);
+		}
+		return null;
+	}
+
+	Future<void> _applyFiltersAndSort() async {
+		List<ContactEntry> filtered =
+				(await SearchService.searchEntriesWithOcr(allImages, searchQuery))
+						.where((img) {
+			final tag = img.state ?? path.basename(path.dirname(img.imagePath));
+			final matchesState = selectedState == 'All' || tag == selectedState;
+			return matchesState;
+		}).toList();
+
+		int compare(ContactEntry a, ContactEntry b) {
+			int result;
+			switch (selectedSortOption) {
+				case 'Date found':
+					result = a.dateFound.compareTo(b.dateFound);
+					break;
+				case 'Size':
+					result = File(a.imagePath)
+							.lengthSync()
+							.compareTo(File(b.imagePath).lengthSync());
+					break;
+				case 'Snap Added Date':
+					DateTime aDate =
+							a.dateAddedOnSnap ?? DateTime.fromMillisecondsSinceEpoch(0);
+					DateTime bDate =
+							b.dateAddedOnSnap ?? DateTime.fromMillisecondsSinceEpoch(0);
+					result = aDate.compareTo(bDate);
+					break;
+				case 'Instagram Added Date':
+					DateTime aDate =
+							a.dateAddedOnInsta ?? DateTime.fromMillisecondsSinceEpoch(0);
+					DateTime bDate =
+							b.dateAddedOnInsta ?? DateTime.fromMillisecondsSinceEpoch(0);
+					result = aDate.compareTo(bDate);
+					break;
+				case 'Added on Snapchat':
+					result = (a.addedOnSnap ? 1 : 0).compareTo(b.addedOnSnap ? 1 : 0);
+					break;
+				case 'Added on Instagram':
+					result = (a.addedOnInsta ? 1 : 0).compareTo(b.addedOnInsta ? 1 : 0);
+					break;
+				case 'Name':
+				default:
+					result =
+							path.basename(a.imagePath).compareTo(path.basename(b.imagePath));
+			}
+			return isAscending ? result : -result;
+		}
+
+		filtered.sort(compare);
+
+		setState(() {
+			images = filtered;
+			currentIndex = 0;
+		});
+		_pageController.jumpToPage(0);
+	}
+
+	Future<String?> _selectState(String currentState) async {
+		String? selected = currentState.isNotEmpty
+				? currentState
+				: states.firstWhere(
+						(s) => s != 'All',
+						orElse: () => '',
+					);
+		final controller = TextEditingController(text: selected);
+
+		final dialogCtx = _navigatorKey.currentContext ?? context;
+		return showDialog<String>(
+			context: dialogCtx,
+			builder: (context) {
+				return StatefulBuilder(
+					builder: (context, setState) {
+						return AlertDialog(
+							title: const Text('Move to state'),
+							content: SingleChildScrollView(
+								child: Column(
+									mainAxisSize: MainAxisSize.min,
+									children: [
+										if (currentState.isNotEmpty)
+											Padding(
+												padding: const EdgeInsets.only(bottom: 8.0),
+												child: Row(
+													children: [
+														const Text('Current:'),
+														const SizedBox(width: 4),
+														Chip(
+															label: Text(currentState),
+															backgroundColor: Colors.grey.shade300,
+														),
+													],
+												),
+											),
+										if (states.where((s) => s != 'All').isNotEmpty)
+											Wrap(
+												spacing: 6,
+												children: [
+													...states.where((s) => s != 'All').map(
+																(s) => ChoiceChip(
+																	label: Text(s),
+																	selected: selected == s,
+																	onSelected: (_) {
+																		setState(() {
+																			selected = s;
+																			controller.text = s;
+																		});
+																	},
+																),
+															),
+													const SizedBox(height: 8),
+												],
+											),
+										TextField(
+											controller: controller,
+											decoration: const InputDecoration(labelText: 'State'),
+										),
+									],
+								),
+							),
+							actions: [
+								TextButton(
+									onPressed: () => Navigator.pop(context),
+									child: const Text('Cancel'),
+								),
+								TextButton(
+									onPressed: () =>
+											Navigator.pop(context, controller.text.trim()),
+									child: const Text('OK'),
+								),
+							],
+						);
+					},
+				);
+			},
+		);
+	}
+
+	Future<void> _onMenuOptionSelected(String imagePath, String option) async {
+		if (option != 'move') return;
+		final targets = <ContactEntry>[];
+
+		if (selectedImages.isNotEmpty) {
+			for (final id in selectedImages) {
+				final match = allImages.where((e) => e.identifier == id);
+				if (match.isNotEmpty) targets.add(match.first);
+			}
+		} else {
+			final match = allImages.where((e) => e.imagePath == imagePath);
+			if (match.isNotEmpty) targets.add(match.first);
+		}
+
+		if (targets.isEmpty) return;
+
+		String currentState = '';
+		if (targets.isNotEmpty) {
+			final firstState = targets.first.state ?? '';
+			if (targets.every((e) => e.state == firstState)) {
+				currentState = firstState;
+			}
+		}
+
+		final newState = await _selectState(currentState);
+		if (newState == null || newState.isEmpty) return;
+
+		setState(() {
+			for (final entry in targets) {
+				entry.state = newState;
+			}
+			selectedImages.clear();
+		});
+
+		_updateStates(allImages);
+		await _applyFiltersAndSort();
+	}
+
+	Future<void> _importImages(BuildContext pickerContext) async {
+		if (_importDirPath == null) {
+			await _changeImportDir();
+			if (_importDirPath == null) {
+				ScaffoldMessenger.of(pickerContext).showSnackBar(
+					const SnackBar(content: Text('No directory selected')),
+				);
+				return;
+			}
+		}
+		final filter = await _buildImportFilter();
+		final ps = await PhotoManager.requestPermissionExtend();
+		if (ps != PermissionState.authorized && ps != PermissionState.limited) {
+			ScaffoldMessenger.of(pickerContext).showSnackBar(
+				const SnackBar(content: Text('Permission not granted')),
+			);
+			return;
+		}
+
+		final config = AssetPickerConfig(
+			requestType: RequestType.image,
+			filterOptions: filter,
+			textDelegate: const EnglishAssetPickerTextDelegate(),
+		);
+
+		final List<AssetEntity>? assets = await AssetPicker.pickAssets(
+			pickerContext,
+			pickerConfig: config,
+		);
+		if (assets == null || assets.isEmpty) {
+			ScaffoldMessenger.of(pickerContext).showSnackBar(
+				const SnackBar(content: Text('No images selected')),
+			);
+			return;
+		}
+
+		final destDir = Directory('/storage/emulated/0/DCIM/Comb');
+		await destDir.create(recursive: true);
+		final messenger = ScaffoldMessenger.of(pickerContext);
+		final total = assets.length;
+		messenger.showSnackBar(
+			SnackBar(
+				duration: const Duration(hours: 1),
+				content: Row(
+					children: [
+						const SizedBox(
+							width: 20,
+							height: 20,
+							child: CircularProgressIndicator(),
+						),
+						const SizedBox(width: 16),
+						Text('Importing images... 0/$total'),
+					],
+				),
+			),
+		);
+
+		List<ContactEntry> newEntries = [];
+		var processed = 0;
+
+		for (final asset in assets) {
+			try {
+				final origin = await asset.originFile;
+				if (origin == null) continue;
+
+				final filename = path.basename(origin.path);
+				final destPath = path.join(destDir.path, filename);
+
+				if (File(destPath).existsSync()) {
+					debugPrint('File already exists at $destPath, skipping');
+					continue;
+				}
+
+				try {
+					await origin.rename(destPath);
+				} catch (_) {
+					// Fall back to copy/delete if rename fails due to SAF restrictions
+					await origin.copy(destPath);
+					try {
+						await origin.delete();
+					} catch (_) {}
+				}
+
+				final id = path.basenameWithoutExtension(filename);
+				final entry = ContactEntry(
+					identifier: id,
+					imagePath: destPath,
+					dateFound: File(destPath).lastModifiedSync(),
+					json: {SubKeys.State: 'Comb'},
+				);
+
+				final result =
+						await ChatGPTService.processImage(imageFile: File(destPath));
+				if (result != null) {
+					postProcessChatGptResult(entry, result, save: false);
+				}
+
+				await StorageUtils.save(entry, backup: false);
+				StorageUtils.filePaths[id] = destPath;
+				newEntries.add(entry);
+			} catch (e) {
+				debugPrint('Failed to import ${asset.id}: $e');
+			}
+
+			processed++;
+			messenger
+				..hideCurrentSnackBar()
+				..showSnackBar(
+					SnackBar(
+						duration: const Duration(hours: 1),
+						content: Row(
+							children: [
+								const SizedBox(
+									width: 20,
+									height: 20,
+									child: CircularProgressIndicator(),
+								),
+								const SizedBox(width: 16),
+								Text('Importing images... $processed/$total'),
+							],
+						),
+					),
+				);
+		}
+
+		await StorageUtils.writeJson(StorageUtils.filePaths);
+
+		if (newEntries.isNotEmpty) {
+			setState(() {
+				allImages.addAll(newEntries);
+			});
+			_updateStates(allImages);
+			await _applyFiltersAndSort();
+			await StorageUtils.syncLocalAndCloud();
+		}
+
+		messenger.hideCurrentSnackBar();
+		final imported = newEntries.length;
+		messenger.showSnackBar(
+			SnackBar(
+				content: Text(
+					imported > 0
+							? 'Imported $imported image${imported == 1 ? '' : 's'}'
+							: 'No new images imported',
+				),
+			),
+		);
+	}
+}
+
+// Updated ImageGallery Widget
+class ImageGallery extends StatelessWidget {
+	final List<ContactEntry> images;
+	final List<String> selectedImages;
+	final Function(String) onImageSelected;
+	final Function(String, String) onMenuOptionSelected;
+	final double galleryHeight; // New dynamic height parameter
+	final ValueChanged<int> onPageChanged;
+	final int currentIndex;
+	final String sortOption;
+
+	ImageGallery({
+		required this.images,
+		required this.selectedImages,
+		required this.onImageSelected,
+		required this.onMenuOptionSelected,
+		required this.galleryHeight,
+		required this.onPageChanged,
+		required this.currentIndex,
+		required this.sortOption,
+	});
+
+	@override
+	Widget build(BuildContext context) {
+		return SizedBox(
+			height: galleryHeight,
+			child: Stack(
+				children: [
+						PageView.builder(
+							controller: _pageController,
+							itemCount: images.length,
+							onPageChanged: onPageChanged,
+							itemBuilder: (context, index) {
+								return ImageTile(
+									imagePath: images[index].imagePath,
+									isSelected:
+											selectedImages.contains(images[index].identifier),
+									extractedText: images[index].extractedText ?? "",
+									identifier: images[index].identifier,
+									sortOption: sortOption,
+									onSelected: onImageSelected,
+									onMenuOptionSelected: onMenuOptionSelected,
+									contact: images[index],
+								);
+							},
+						),
+					Positioned(
+						bottom: 8,
+						left: 0,
+						right: 0,
+						child: Center(
+							child: Container(
+								padding: EdgeInsets.symmetric(vertical: 4, horizontal: 8),
+								decoration: BoxDecoration(
+									color: Colors.black54,
+									borderRadius: BorderRadius.circular(8),
+								),
+								child: Text(
+									'${currentIndex + 1} / ${images.length}',
+									style: TextStyle(color: Colors.white),
+								),
+							),
+						),
+					),
+				],
+			),
+		);
+	}
+}
+
+// Updated ImageTile Widget with Selection
+class ImageTile extends StatefulWidget {
+	final String imagePath;
+	final bool isSelected;
+	final String extractedText;
+	final String identifier;
+	final String sortOption;
+	final Function(String) onSelected;
+	final Function(String, String) onMenuOptionSelected;
+	final ContactEntry contact;
+
+	ImageTile({
+		required this.imagePath,
+		required this.isSelected,
+		required this.extractedText,
+		required this.identifier,
+		required this.sortOption,
+		required this.onSelected,
+		required this.onMenuOptionSelected,
+		required this.contact,
+	});
+
+	@override
+	_ImageTileState createState() => _ImageTileState();
+}
+
+class _ImageTileState extends State<ImageTile> {
+	String get _truncatedText {
+		const maxChars = 120;
+		if (widget.extractedText.length <= maxChars) return widget.extractedText;
+		return '${widget.extractedText.substring(0, maxChars)}...';
+	}
+
+	String get _displayLabel {
+		switch (widget.sortOption) {
+			case 'Date found':
+				return DateFormat.yMd().format(widget.contact.dateFound);
+			case 'Snap Added Date':
+				final snapDate = widget.contact.dateAddedOnSnap;
+				return snapDate != null ? DateFormat.yMd().format(snapDate) : 'No date';
+			case 'Instagram Added Date':
+				final instaDate = widget.contact.dateAddedOnInsta;
+				return instaDate != null
+						? DateFormat.yMd().format(instaDate)
+						: 'No date';
+			case 'Added on Snapchat':
+				return widget.contact.addedOnSnap ? 'Added' : 'Not Added';
+			case 'Added on Instagram':
+				return widget.contact.addedOnInsta ? 'Added' : 'Not Added';
+			case 'Name':
+				return widget.contact.name ?? widget.identifier;
+			default:
+				return widget.identifier;
+		}
+	}
+
+	// Move all helper methods above build
+	void _showPopupMenu(BuildContext context, String imagePath) {
+		showModalBottomSheet(
+			context: context,
+			builder: (sheetContext) {
+				return ListView(
+					shrinkWrap: true,
+					children: [
+						ListTile(
+							leading: Icon(Icons.info),
+							title: Text('View Details'),
+							onTap: () {
+								Navigator.pop(sheetContext);
+								_showDetailsDialog(context);
+							},
+						),
+						ListTile(
+							leading: Icon(Icons.refresh),
+							title: Text('Redo text extraction'),
+							onTap: () async {
+								Navigator.pop(sheetContext);
+								await _redoTextExtraction();
+							},
+						),
+						if (widget.contact.snapUsername?.isNotEmpty ?? false)
+							ListTile(
+								leading: Icon(Icons.chat_bubble),
+								title: Text('Open on Snap'),
+								onTap: () {
+									Navigator.pop(sheetContext);
+									_openSocial(
+											SocialType.Snapchat, widget.contact.snapUsername!);
+								},
+							),
+						if (widget.contact.instaUsername?.isNotEmpty ?? false)
+							ListTile(
+								leading: Icon(Icons.camera_alt),
+								title: Text('Open on Insta'),
+								onTap: () {
+									Navigator.pop(sheetContext);
+									_openSocial(
+											SocialType.Instagram, widget.contact.instaUsername!);
+								},
+							),
+						if (widget.contact.discordUsername?.isNotEmpty ?? false)
+							ListTile(
+								leading: Icon(Icons.discord),
+								title: Text('Open on Discord'),
+								onTap: () {
+									Navigator.pop(sheetContext);
+									_openSocial(
+											SocialType.Discord, widget.contact.discordUsername!);
+								},
+							),
+						if (widget.contact.addedOnSnap) ...[
+							ListTile(
+								leading: Icon(Icons.person_off),
+								title: Text('Quick Unfriend on Snap'),
+								onTap: () async {
+									Navigator.pop(sheetContext);
+									await _unfriendSocial(
+										SocialType.Snapchat,
+										autoReason: 'no response',
+									);
+								},
+							),
+							ListTile(
+								leading: Icon(Icons.person_remove),
+								title: Text('Unfriend on Snap with Note'),
+								onTap: () async {
+									Navigator.pop(sheetContext);
+									await _unfriendSocial(
+										SocialType.Snapchat,
+										promptForNote: true,
+										autoReason: 'conversation ended',
+									);
+								},
+							),
+						],
+						if (widget.contact.addedOnInsta) ...[
+							ListTile(
+								leading: Icon(Icons.person_off),
+								title: Text('Quick Unfriend on Insta'),
+								onTap: () async {
+									Navigator.pop(sheetContext);
+									await _unfriendSocial(
+										SocialType.Instagram,
+										autoReason: 'no response',
+									);
+								},
+							),
+							ListTile(
+								leading: Icon(Icons.person_remove),
+								title: Text('Unfriend on Insta with Note'),
+								onTap: () async {
+									Navigator.pop(sheetContext);
+									await _unfriendSocial(
+										SocialType.Instagram,
+										promptForNote: true,
+										autoReason: 'conversation ended',
+									);
+								},
+							),
+						],
+						if (widget.contact.addedOnDiscord) ...[
+							ListTile(
+								leading: Icon(Icons.person_off),
+								title: Text('Quick Unfriend on Discord'),
+								onTap: () async {
+									Navigator.pop(sheetContext);
+									await _unfriendSocial(
+										SocialType.Discord,
+										autoReason: 'no response',
+									);
+								},
+							),
+							ListTile(
+								leading: Icon(Icons.person_remove),
+								title: Text('Unfriend on Discord with Note'),
+								onTap: () async {
+									Navigator.pop(sheetContext);
+									await _unfriendSocial(
+										SocialType.Discord,
+										promptForNote: true,
+										autoReason: 'conversation ended',
+									);
+								},
+							),
+						],
+						ListTile(
+							leading: Icon(Icons.note_alt_outlined),
+							title: Text('Edit Notes'),
+							onTap: () async {
+								Navigator.pop(sheetContext);
+								await showNoteDialog(
+										context, widget.contact.identifier, widget.contact,
+										existingNotes: widget.contact.notes);
+							},
+						),
+						ListTile(
+							leading: Icon(Icons.move_to_inbox),
+							title: Text('Move'),
+							onTap: () {
+								widget.onMenuOptionSelected(imagePath, 'move');
+								Navigator.pop(sheetContext);
+							},
+						),
+					],
+				);
+			},
+		);
+	}
+
+	Future<void> _redoTextExtraction() async {
+		final result = await Navigator.of(context).push<Map<String, dynamic>>(
+			MaterialPageRoute(
+				builder: (_) => RedoCropScreen(imageFile: File(widget.imagePath)),
+			),
+		);
+		if (result != null) {
+			setState(() {
+				postProcessChatGptResult(widget.contact, result, save: false);
+			});
+			await StorageUtils.save(widget.contact, backup: false);
+		}
+	}
+
+	void _showDetailsDialog(BuildContext context) {
+		showDialog(
+			context: context,
+			builder: (_) => Dialog(
+				child: Column(
+					mainAxisSize: MainAxisSize.min,
+					children: [
+						Container(
+							constraints: BoxConstraints(maxHeight: 300),
+							child: PhotoView(
+								imageProvider: FileImage(File(widget.imagePath)),
+								backgroundDecoration: BoxDecoration(color: Colors.white),
+								minScale: PhotoViewComputedScale.contained,
+								maxScale: PhotoViewComputedScale.covered * 2.5,
+							),
+						),
+						Padding(
+							padding: const EdgeInsets.all(12.0),
+							child: Text(
+								widget.extractedText.isNotEmpty
+										? widget.extractedText
+										: 'No text found',
+							),
+						),
+						TextButton(
+							onPressed: () => Navigator.pop(context),
+							child: Text('Close'),
+						),
+					],
+				),
+			),
+		);
+	}
+
+	Future<bool> _confirm(BuildContext context,
+			{String message = 'Are you sure?'}) async {
+		final result = await showDialog<bool>(
+			context: context,
+			builder: (_) => ConfirmationDialog(message: message),
+		);
+		return result ?? false;
+	}
+
+	Future<void> _editUsernames(BuildContext context) async {
+		final originalSnap = widget.contact.snapUsername ?? '';
+		final originalInsta = widget.contact.instaUsername ?? '';
+		final originalDiscord = widget.contact.discordUsername ?? '';
+
+		final snapController = TextEditingController(text: originalSnap);
+		final instaController = TextEditingController(text: originalInsta);
+		final discordController = TextEditingController(text: originalDiscord);
+
+		bool changed = false;
+		void updateChanged() {
+			changed = snapController.text != originalSnap ||
+					instaController.text != originalInsta ||
+					discordController.text != originalDiscord;
+		}
+
+		snapController.addListener(updateChanged);
+		instaController.addListener(updateChanged);
+		discordController.addListener(updateChanged);
+
+		final result = await showDialog<List<String>>(
+			context: context,
+			barrierDismissible: false,
+			builder: (context) {
+				return WillPopScope(
+					onWillPop: () async {
+						if (changed) {
+							return await _confirm(context, message: 'Discard changes?');
+						}
+						return true;
+					},
+					child: AlertDialog(
+						title: Text('Edit Usernames'),
+						content: Column(
+							mainAxisSize: MainAxisSize.min,
+							children: [
+								TextField(
+									controller: snapController,
+									decoration: InputDecoration(labelText: 'Snapchat'),
+								),
+								TextField(
+									controller: instaController,
+									decoration: InputDecoration(labelText: 'Instagram'),
+								),
+								TextField(
+									controller: discordController,
+									decoration: InputDecoration(labelText: 'Discord'),
+								),
+							],
+						),
+						actions: [
+							TextButton(
+								onPressed: () async {
+									if (changed) {
+										final discard =
+												await _confirm(context, message: 'Discard changes?');
+										if (!discard) return;
+									}
+									Navigator.pop(context);
+								},
+								child: Text('Cancel'),
+							),
+							TextButton(
+								onPressed: () async {
+									if (changed) {
+										final confirmSave =
+												await _confirm(context, message: 'Save changes?');
+										if (!confirmSave) return;
+									}
+									Navigator.pop(context, [
+										snapController.text,
+										instaController.text,
+										discordController.text,
+									]);
+								},
+								child: Text('Save'),
+							),
+						],
+					),
+				);
+			},
+		);
+
+		if (result != null) {
+			if (result[0] != widget.contact.snapUsername) {
+				await SocialType.Snapchat.saveUsername(widget.contact, result[0],
+						overriding: true);
+			}
+			if (result[1] != widget.contact.instaUsername) {
+				await SocialType.Instagram.saveUsername(widget.contact, result[1],
+						overriding: true);
+			}
+			if (result[2] != widget.contact.discordUsername) {
+				await SocialType.Discord.saveUsername(widget.contact, result[2],
+						overriding: true);
+			}
+			setState(() {});
+		}
+	}
+
+	void _openSocial(SocialType social, String username) async {
+		Uri url;
+		switch (social) {
+			case SocialType.Snapchat:
+				url = Uri.parse('https://www.snapchat.com/add/${username.toLowerCase()}');
+				break;
+			case SocialType.Instagram:
+				url = Uri.parse('https://www.instagram.com/$username');
+				break;
+			case SocialType.Discord:
+				Clipboard.setData(ClipboardData(text: username));
+				SocialIcon.discordIconButton?.openApp();
+				return;
+			default:
+				return;
+		}
+		if (!url.hasEmptyPath) {
+			await launchUrl(url, mode: LaunchMode.externalApplication);
+		}
+	}
+
+	Future<void> _unfriendSocial(
+		SocialType social, {
+		bool promptForNote = false,
+		String? autoReason,
+	}) async {
+		final username = await social.getUserName(widget.contact);
+		if (username == null || username.isEmpty) return;
+
+		_openSocial(social, username);
+		bool res = await _confirm(context, message: 'Confirm unfriended?');
+		if (!res) return;
+
+		setState(() {
+			widget.contact.state = 'Strings';
+
+			final now = DateFormat.yMd().add_jm().format(DateTime.now());
+			String note = 'Unfriended from ${social.name} on $now';
+			if (autoReason != null && autoReason.isNotEmpty) {
+				note += ' ($autoReason)';
+			}
+			if (widget.contact.notes == null || widget.contact.notes!.isEmpty) {
+				widget.contact.notes = note;
+			} else {
+				widget.contact.notes = '${widget.contact.notes}\n$note';
+			}
+		});
+
+		if (promptForNote) {
+			final extra = await showNoteDialog(
+				context,
+				widget.contact.identifier,
+				widget.contact,
+			);
+			if (extra != null && extra.isNotEmpty) {
+				setState(() {
+					widget.contact.notes = '${widget.contact.notes}\n$extra';
+				});
+			}
+		}
+	}
+
+	@override
+	Widget build(BuildContext context) {
+		return GestureDetector(
+			onTap: () => _showDetailsDialog(context),
+			onLongPress: () => widget.onSelected(widget.identifier),
+			child: LayoutBuilder(
+				builder: (context, constraints) {
+					return Container(
+						margin: const EdgeInsets.symmetric(horizontal: 10, vertical: 6),
+						width: constraints.maxWidth * 0.8,
+						decoration: BoxDecoration(
+							borderRadius: BorderRadius.circular(16),
+							border: widget.isSelected
+									? Border.all(color: Colors.blueAccent, width: 3)
+									: null,
+							boxShadow: const [
+								BoxShadow(
+									color: Colors.black26,
+									blurRadius: 6,
+									offset: Offset(0, 4),
+								),
+							],
+						),
+						child: ClipRRect(
+							borderRadius: BorderRadius.circular(16),
+							child: Stack(
+								children: [
+									Positioned.fill(
+										child: PhotoView(
+											imageProvider: FileImage(File(widget.imagePath)),
+											backgroundDecoration:
+													const BoxDecoration(color: Colors.white),
+											minScale: PhotoViewComputedScale.contained,
+											maxScale: PhotoViewComputedScale.covered * 2.5,
+										),
+									),
+									// Redo OCR button overlay
+									Positioned(
+										top: 8,
+										right: 8,
+										child: Material(
+											color: Colors.transparent,
+											shape: const CircleBorder(),
+											elevation: 4,
+											child: InkWell(
+												customBorder: const CircleBorder(),
+												onTap: () async {
+													await _redoTextExtraction();
+												},
+												child: Container(
+													width: 36,
+													height: 36,
+													decoration: BoxDecoration(
+														shape: BoxShape.circle,
+														gradient: const LinearGradient(
+															colors: [Color(0xFF4F8CFF), Color(0xFF8F5CFF)],
+															begin: Alignment.topLeft,
+															end: Alignment.bottomRight,
+														),
+														border: Border.all(color: Colors.white, width: 2),
+														boxShadow: [
+															BoxShadow(
+																color: Colors.black26,
+																blurRadius: 8,
+																offset: Offset(0, 2),
+															),
+														],
+													),
+													child: const Center(
+														child: Icon(
+															Icons.refresh,
+															color: Colors.white,
+															size: 20,
+														),
+													),
+												),
+											),
+										),
+									),
+									// Existing info and selection UI
+									Positioned(
+										top: 8,
+										left: 8,
+										child: Column(
+											crossAxisAlignment: CrossAxisAlignment.start,
+											children: [
+												ConstrainedBox(
+													constraints: BoxConstraints(
+															maxWidth: constraints.maxWidth - 50),
+													child: Container(
+														padding: const EdgeInsets.symmetric(
+																vertical: 2, horizontal: 6),
+														decoration: BoxDecoration(
+															color: Colors.black54,
+															borderRadius: BorderRadius.circular(6),
+														),
+														child: Text(
+															_displayLabel,
+															maxLines: 1,
+															overflow: TextOverflow.ellipsis,
+															style: const TextStyle(
+																color: Colors.white,
+																fontSize: 12,
+																fontWeight: FontWeight.bold,
+															),
+														),
+													),
+												),
+												const SizedBox(height: 4),
+												GestureDetector(
+													onTap: () => widget.onSelected(widget.identifier),
+													child: Icon(
+														widget.isSelected
+																? Icons.check_circle
+																: Icons.radio_button_unchecked,
+														color: widget.isSelected
+																? Colors.blueAccent
+																: Colors.grey,
+														size: 28,
+													),
+												),
+											],
+										),
+									),
+									Positioned(
+										left: 0,
+										right: 0,
+										bottom: 0,
+										child: Container(
+											height: 56,
+											padding: const EdgeInsets.symmetric(horizontal: 8),
+											decoration: const BoxDecoration(
+												gradient: LinearGradient(
+													begin: Alignment.topCenter,
+													end: Alignment.bottomCenter,
+													colors: [Colors.transparent, Colors.black54],
+												),
+											),
+											child: Row(
+												children: [
+													Text(
+														_truncatedText,
+														maxLines: 1,
+														overflow: TextOverflow.ellipsis,
+														style: const TextStyle(color: Colors.white),
+													),
+													Flexible(
+														child: SingleChildScrollView(
+															scrollDirection: Axis.horizontal,
+															child: Row(
+																children: [
+																	if (widget.contact.snapUsername?.isNotEmpty ??
+																			false)
+																		IconButton(
+																			iconSize: 22,
+																			color: Colors.white,
+																			padding: EdgeInsets.zero,
+																			constraints:
+																					const BoxConstraints.tightFor(
+																							width: 36, height: 36),
+																			onPressed: () => _openSocial(
+																					SocialType.Snapchat,
+																					widget.contact.snapUsername!),
+																			icon: SocialIcon
+																					.snapchatIconButton!.socialIcon,
+																		),
+																	if (widget
+																					.contact.instaUsername?.isNotEmpty ??
+																			false)
+																		IconButton(
+																			iconSize: 22,
+																			color: Colors.white,
+																			padding: EdgeInsets.zero,
+																			constraints:
+																					const BoxConstraints.tightFor(
+																							width: 36, height: 36),
+																			onPressed: () => _openSocial(
+																					SocialType.Instagram,
+																					widget.contact.instaUsername!),
+																			icon: SocialIcon
+																					.instagramIconButton!.socialIcon,
+																		),
+																	if (widget.contact.discordUsername
+																					?.isNotEmpty ??
+																			false)
+																		IconButton(
+																			iconSize: 22,
+																			color: Colors.white,
+																			padding: EdgeInsets.zero,
+																			constraints:
+																					const BoxConstraints.tightFor(
+																							width: 36, height: 36),
+																			onPressed: () => _openSocial(
+																					SocialType.Discord,
+																					widget.contact.discordUsername!),
+																			icon: SocialIcon
+																					.discordIconButton!.socialIcon,
+																		),
+																	IconButton(
+																		iconSize: 22,
+																		color: Colors.white,
+																		padding: EdgeInsets.zero,
+																		constraints: const BoxConstraints.tightFor(
+																				width: 36, height: 36),
+																		icon: const Icon(Icons.note_alt_outlined),
+																		onPressed: () async {
+																			await showNoteDialog(
+																					context,
+																					widget.contact.identifier,
+																					widget.contact,
+																					existingNotes: widget.contact.notes);
+																		},
+																	),
+																	IconButton(
+																		iconSize: 22,
+																		color: Colors.white,
+																		padding: EdgeInsets.zero,
+																		constraints: const BoxConstraints.tightFor(
+																				width: 36, height: 36),
+																		icon: const Icon(Icons.edit),
+																		onPressed: () => _editUsernames(context),
+																	),
+																	IconButton(
+																		iconSize: 22,
+																		color: Colors.white,
+																		padding: EdgeInsets.zero,
+																		constraints: const BoxConstraints.tightFor(
+																				width: 36, height: 36),
+																		icon: const Icon(Icons.more_vert),
+																		onPressed: () => _showPopupMenu(
+																				context, widget.imagePath),
+																	),
+																],
+															),
+														),
+													),
+												],
+											),
+										),
+									),
+								],
+							),
+						),
+					);
+				},
+			),
+		);
+	}
+}
+
